moved the logic to graph window class
